finish car and road grount truth processor

+++ car_road_ground_truth.py
@@ -20,25 +20,28 @@ if not os.path.isdir(seg_folder):
     sys.exit()
 
 car_seg_folder = os.path.dirname(seg_folder) + '/' + car_seg_folder
-shutil.rmtree(car_seg_folder)
+if os.path.exists(car_seg_folder):
+    shutil.rmtree(car_seg_folder)
 os.makedirs(car_seg_folder)
+
 road_seg_folder = os.path.dirname(seg_folder) + '/' + road_seg_folder
-shutil.rmtree(road_seg_folder)
+if os.path.exists(road_seg_folder):
+    shutil.rmtree(road_seg_folder)
 os.makedirs(road_seg_folder)
 
 image_files = glob.glob(seg_folder + '/*.png')
 for image_file in image_files:
     image = cv2.cvtColor(cv2.imread(image_file), cv2.COLOR_BGR2RGB)
+    # Ignore car pixels after row 490
+    red490 = image[490:,:,0]
+    red490[red490 == 10] = 0
     # Grab red channel
-    red = image[:,:,0]
-    red = np.expand_dims(red, axis=2)
+    red = image[:,:,0,np.newaxis]
     # Look for cars
-    car_image = np.where((red == 10).all(), np.ones(red.shape[:-1] + (3,)) * 255, np.ones(red.shape[:-1] + (3,)))
-    #print(car_image.shape)
-    #print((np.ones(red.shape[:-1] + (3,)) * 255).shape)
+    car_image = np.where(red == 10, [255, 255, 255], [0, 0, 0])
     # Save car image
     cv2.imwrite(car_seg_folder + '/' + os.path.basename(image_file), car_image)
     # Look for roads
-    road_image = np.where((red == 6).all() or (red == 7).all(), np.ones(red.shape[:-1] + (3,)) * 255, np.ones(red.shape[:-1] + (3,)))
+    road_image = np.where(np.logical_or(red == 6, red == 7), [255, 255, 255], [0, 0, 0])
     # Save road image
     cv2.imwrite(road_seg_folder + '/' + os.path.basename(image_file), road_image)
